+++ lib/kernels/src/cpu/reverse_kernels.cc
@@ -11,13 +11,13 @@ struct CPUReverseForwardKernel {
                   GenericTensorAccessorW &output) {
     assert(input.data_type == DT && output.data_type == DT);
 
-    size_t num_out_blocks = input.shape.at(legion_dim_t(0));
-    size_t reverse_dim_size = input.shape.at(legion_dim_t(1));
-    size_t in_block_size = input.shape.at(legion_dim_t(2));
+    int num_out_blocks = input.shape.at(legion_dim_t(0));
+    int reverse_dim_size = input.shape.at(legion_dim_t(1));
+    int in_block_size = input.shape.at(legion_dim_t(2));
 
-    for (size_t block_idx = 0; block_idx < num_out_blocks; block_idx++) {
-      for (size_t rev_idx = 0; rev_idx < reverse_dim_size; rev_idx++) {
-        for (size_t i = 0; i < in_block_size; i++) {
+    for (int block_idx = 0; block_idx < num_out_blocks; block_idx++) {
+      for (int rev_idx = 0; rev_idx < reverse_dim_size; rev_idx++) {
+        for (int i = 0; i < in_block_size; i++) {
           output.at<DT>({block_idx, rev_idx, i}) =
               input.at<DT>({num_out_blocks - 1 - block_idx,
                             reverse_dim_size - 1 - rev_idx,
